disable seek-with-arrow-key control by native <video> widget
With a keydown event handler that triggers during the capturing
phase, disable the default action performed by the <video> widget
for ArrowLeft and ArrowRight keys.  This allows our event handler
to handle these keys to seek the video without getting affected by
what the native <video> does in addition to what we do.