playlist: fix moving non-contiguous selection downwards
When a non-contiguous set of items in the playlist is selected,
pressing Ctrl-Up-Arrow or Ctrl-Down-Arrow gathers the selected
items. It does this by moving the unselected items from the
selection span to above or below the top of the selection.

In the current implementation, the relative order of these moved
items is inadvertently reversed upon Ctrol-Down-Arrow.

This behavior is inconsistent with the Ctrl-Up-Arrow action, which
preserves the relative order of the intermediate items it
moves. This inconsistency can be confusing to the user.

To fix this, process the array of non-selected items in its natural
order instead of a reversed one. This makes the downward-gathering
behavior consistent with the upward-gathering one.